[#15] feat: HistoryLog Enum 타입 파일 생성 및 분리
