+++ ProjectManager/ProjectManager/Model/HistoryManager.swift
@@ -1,22 +1,5 @@
 import Foundation
 
-enum HistoryLog {
-    case add(String)
-    case move(String, String, String)
-    case delete(String, String)
-    
-    var description: String {
-        switch self {
-        case .add(let title):
-            return "Added '\(title)'."
-        case .move(let title, let before,  let after):
-            return "Moved '\(title)' from \(before) to \(after)."
-        case .delete(let title, let before):
-            return "Removed '\(title)' from \(before)."
-        }
-    }
-}
-
 class HistoryManager {
     static let shared = HistoryManager()
